chore: simplify app navigator
